Add a button and action to "reselect" the selection rect
There is no way to check when IKImageView finished its image rendering process and after it's done the selection rect is usually left behind the rendered image.
The current code tries to reselect it after 1 second delay, but sometimes this doesn't work, so as a workaround the button to reset the selection rect manually.