Use dispatch_sync instead of CFRunLoopPerformBlock
Since the CFRunLoopPerformBlock() function is asynchronous the state of bound buttons is not updated properly.
Everything is normal for a single-pass encoding, but if a two-pass encoding is in progress, then during the second pass
the state of the bound control buttons is not correctly set. The dispatch_sync() function fixes this problem.