Add NSDictionary category for compatibility with older systems
The old "unsafe" methods to read and write dictionaries as XML files are deprecated, and the new ones are not available on pre 10.13 systems.
This category provides an implementation for "safe" read/write methods on older systems. The category is complied only if the deployment target is set to 10.12 or lower, otherwise Apple's methods from Foundation framework are used instead.