Use latest macOS SDK by default
Since IOSurface-based NSView is much stable on macOS 10.14 than NSOpenGLView, the project can be build with the latest SDK now.
Old 10.13 SDK was used previously for compatibility reasons, because NSOpenGLView appears to be more stable on macOS 10.14 when compiled with the old SDK, but it's messed up if 10.14+ SDK is used.